renamed SizeField, edited SliderGradient
